Continuing merging from array-rework branch. Still not compiling.. and still incomplete...
git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/FreeMat4@3340 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ libs/libXP/QTTerm.cpp
@@ -422,7 +422,7 @@ void QTTerm::keyPressEvent(QKeyEvent *e) {
 }
 
 void QTTerm::OutputRawString(QString txt) {
-  for (size_t i=0;i<txt.size();i++)
+  for (int i=0;i<txt.size();i++)
     setChar(txt[i].toAscii(),true);
 }
